requirements.txt
aniso8601==8.0.0
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
flask-marshmallow==0.9.0
Flask-MySQL==1.4.0
Flask-RESTful==0.3.7
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10.1
lxml==4.4.1
MarkupSafe==1.1.1
marshmallow==3.0.5
mysql-connector-python==8.0.17
newspaper3k==0.2.8
nltk==3.4.5
Pillow==6.1.0
protobuf==3.9.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
six==1.12.0
soupsieve==1.9.3
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.25.3
Werkzeug==0.15.6
